Moved all *ViewModel classes to viewmodel package and renamed ProfileViewModel to LogOutViewModel, LoginViewModel to LogInViewModel and SplashViewModel to CheckExistingSessionViewModel classes
